requirements.txt
apiclient==1.0.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
gTTS==2.3.1
httplib2==0.22.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
moviepy==1.0.3
mutagen==1.46.0
numpy==1.24.2
opencv-python-headless==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
praw==7.7.0
prawcore==2.3.0
proglog==0.1.10
protobuf==4.21.12
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pywatchman==1.4.1
reportlab==3.6.12
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
tqdm==4.65.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
